Huge JS Optimization: Fix Tree Shaking issue caused by AddressCopyButton
