+++ tw-components/button.tsx
@@ -1,6 +1,4 @@
-import { Card } from "./card";
 import { ChakraNextLink } from "./link";
-import { Text } from "./text";
 import { convertFontSizeToCSSVar } from "./utils/typography";
 import {
   Button as ChakraButton,
@@ -10,18 +8,14 @@ import {
   IconButtonProps,
   LightMode,
   Link,
-  Tooltip,
   forwardRef,
   useButtonGroup,
-  useClipboard,
-  useToast,
 } from "@chakra-ui/react";
 import { Link as LocationLink, useMatch } from "@tanstack/react-location";
 import { useTrack } from "hooks/analytics/useTrack";
-import React, { useEffect } from "react";
-import { FiCopy, FiExternalLink } from "react-icons/fi";
+import React from "react";
+import { FiExternalLink } from "react-icons/fi";
 import { fontWeights, letterSpacings, lineHeights } from "theme/typography";
-import { shortenIfAddress } from "utils/usedapp-external";
 
 export const buttonSizesMap = {
   xs: "sm",
@@ -164,83 +158,3 @@ export const TrackedIconButton = forwardRef<TrackedIconButtonProps, "button">(
 );
 
 TrackedIconButton.displayName = "TrackedIconButton";
-
-interface AddressCopyButtonProps extends Omit<ButtonProps, "onClick" | "size"> {
-  address?: string;
-  noIcon?: boolean;
-  size?: PossibleButtonSize;
-  tokenId?: boolean;
-}
-
-export const AddressCopyButton: React.FC<AddressCopyButtonProps> = ({
-  address,
-  noIcon,
-  flexGrow = 0,
-  size = "sm",
-  borderRadius = "md",
-  variant = "outline",
-  tokenId,
-  ...restButtonProps
-}) => {
-  const { onCopy, setValue } = useClipboard(address || "");
-
-  useEffect(() => {
-    if (address) {
-      setValue(address);
-    }
-  }, [address, setValue]);
-
-  const trackEvent = useTrack();
-  const toast = useToast();
-
-  return (
-    <Tooltip
-      p={0}
-      bg="transparent"
-      boxShadow="none"
-      label={
-        <Card py={2} px={4}>
-          <Text size="label.sm">
-            Copy {tokenId ? "Token ID" : "address"} to clipboard
-          </Text>
-        </Card>
-      }
-    >
-      <Button
-        flexGrow={flexGrow}
-        size={size}
-        borderRadius={borderRadius}
-        variant={variant}
-        {...restButtonProps}
-        onClick={(e) => {
-          e.stopPropagation();
-          e.preventDefault();
-          onCopy();
-          toast({
-            variant: "solid",
-            position: "bottom",
-            title: `${tokenId ? "Token ID" : "Address"} copied.`,
-            status: "success",
-            duration: 5000,
-            isClosable: true,
-          });
-          if (tokenId) {
-            trackEvent({
-              category: "tokenid_button",
-              action: "copy",
-              tokenId: address,
-            });
-          } else {
-            trackEvent({ category: "address_button", action: "copy", address });
-          }
-        }}
-        leftIcon={noIcon ? undefined : <Icon boxSize={3} as={FiCopy} />}
-        fontFamily="mono"
-      >
-        <Text size={`label.${buttonSizesMap[size]}`}>
-          {shortenIfAddress(address)}
-        </Text>
-      </Button>
-    </Tooltip>
-  );
-};
