+++ tw-components/index.ts
@@ -9,3 +9,9 @@ export * from "./heading";
 export * from "./text";
 export * from "./form";
 export * from "./menu";
+
+/**
+ * ❗️❗️❗️ @DANGER
+ * Do not export `AddressCopyButton` from here
+ * It will result in a HUGE bundle size
+ */
